refactor(markdown): split formatting helpers
Move low-level Markdown escaping, table-cell rendering, form-field labels, annotation labels, visual-region labels, and JavaScript action formatting out of the formatter body so formatMarkdown can focus on document and page section assembly.
