+++ src/output/markdown.ts
@@ -1,4 +1,28 @@
 import type { DocumentResult, PageResult, PageStructureItem, PageStructureNode } from '../types/index.js';
+import {
+  annotationBorder,
+  annotationColor,
+  annotationFileAttachment,
+  annotationFlags,
+  annotationShape,
+  escapeInline,
+  escapeTableCell,
+  fieldActions,
+  fieldExportValue,
+  fieldFlags,
+  fieldLabel,
+  fieldOptions,
+  fieldResetForm,
+  fieldValue,
+  formatBbox,
+  formatBox,
+  formatJavaScriptActions,
+  formatViewerValue,
+  jsActionCount,
+  linkTarget,
+  visualRegionAssociatedText,
+  visualRegionSources,
+} from './markdown/helpers.js';
 
 /** Options that influence the Markdown rendering without changing the
  *  underlying `DocumentResult`. JSON / XML formatters don't need them
@@ -13,154 +37,10 @@ export interface MarkdownOptions {
   stripRepeated?: boolean;
 }
 
-const MARKDOWN_JS_ACTIONS_MAX_CHARS = 500;
-
 /** "595×842" — drops trailing .00 so integer dimensions stay readable. */
 function formatSize(page: PageResult): string {
   const trim = (n: number) => (Number.isInteger(n) ? String(n) : n.toFixed(2));
   return `${trim(page.width)}×${trim(page.height)}`;
-}
-
-function escapeTableCell(value: string): string {
-  return value
-    .replaceAll('\\', '\\\\')
-    .replaceAll('|', '\\|')
-    .replaceAll('\r\n', ' ')
-    .replaceAll('\n', ' ')
-    .replaceAll('\r', ' ');
-}
-
-function escapeInline(value: string): string {
-  return value
-    .replaceAll('\\', '\\\\')
-    .replaceAll('&', '&amp;')
-    .replaceAll('<', '&lt;')
-    .replaceAll('>', '&gt;')
-    .replaceAll('\n', ' ')
-    .replaceAll('\r', ' ')
-    .replaceAll('`', '\\`')
-    .replaceAll('*', '\\*')
-    .replaceAll('_', '\\_')
-    .replaceAll('[', '\\[')
-    .replaceAll(']', '\\]')
-    .replaceAll('|', '\\|');
-}
-
-function fieldValue(field: NonNullable<PageResult['formFields']>[number]): string {
-  if (field.checked !== undefined) return field.checked ? 'checked' : 'unchecked';
-  if (field.type === 'button' && field.caption) return field.caption;
-  if (field.type === 'choice' && field.displayValue) return field.displayValue;
-  return field.value ?? '';
-}
-
-function fieldLabel(field: NonNullable<PageResult['formFields']>[number]): string {
-  return field.label ? `${field.label.text} (${field.label.relation})` : '';
-}
-
-function fieldOptions(field: NonNullable<PageResult['formFields']>[number]): string {
-  return (
-    field.options
-      ?.map((option) =>
-        option.displayValue === option.exportValue
-          ? option.displayValue
-          : `${option.displayValue}=${option.exportValue}`,
-      )
-      .join(', ') ?? ''
-  );
-}
-
-function fieldExportValue(field: NonNullable<PageResult['formFields']>[number]): string {
-  if (field.type === 'choice' && field.displayValue && field.value && field.displayValue !== field.value) {
-    return field.value;
-  }
-  return field.exportValue ?? '';
-}
-
-function fieldFlags(field: NonNullable<PageResult['formFields']>[number]): string {
-  const flags = new Set<string>(field.flags ?? []);
-  if (field.readOnly) flags.add('readOnly');
-  if (field.required) flags.add('required');
-  if (field.multiline) flags.add('multiline');
-  if (field.combo !== undefined) flags.add(field.combo ? 'combo' : 'list');
-  if (field.multiSelect) flags.add('multiSelect');
-  return Array.from(flags).join(', ');
-}
-
-function fieldActions(field: NonNullable<PageResult['formFields']>[number]): string {
-  return field.actions ? formatJavaScriptActions(field.actions) : '';
-}
-
-function fieldResetForm(field: NonNullable<PageResult['formFields']>[number]): string {
-  if (!field.resetForm) return '';
-  const fields = field.resetForm.fields.join(', ');
-  if (field.resetForm.include) return fields.length > 0 ? `reset only ${fields}` : 'reset only listed fields';
-  return fields.length > 0 ? `reset all except ${fields}` : 'reset all fields';
-}
-
-function annotationColor(annotation: NonNullable<PageResult['annotations']>[number]): string {
-  return annotation.color ? annotation.color.join(',') : '';
-}
-
-function annotationFileAttachment(annotation: NonNullable<PageResult['annotations']>[number]): string {
-  const file = annotation.fileAttachment;
-  if (!file) return '';
-  const parts = [file.name, `${file.size} bytes`];
-  if (file.description) parts.push(file.description);
-  return parts.join(' · ');
-}
-
-function annotationFlags(annotation: NonNullable<PageResult['annotations']>[number]): string {
-  return annotation.flags?.join(',') ?? '';
-}
-
-function annotationBorder(annotation: NonNullable<PageResult['annotations']>[number]): string {
-  const border = annotation.border;
-  if (!border) return '';
-  const parts: string[] = [];
-  if (border.width !== undefined) parts.push(`width=${border.width}`);
-  if (border.style !== undefined) parts.push(border.style);
-  if (border.dashArray !== undefined && border.dashArray.length > 0) parts.push(`dash=${border.dashArray.join(',')}`);
-  return parts.join(' ');
-}
-
-function annotationShape(annotation: NonNullable<PageResult['annotations']>[number]): string {
-  const parts: string[] = [];
-  if (annotation.line) {
-    const { from, to, endings } = annotation.line;
-    const endingText = endings ? ` endings=${endings.join(',')}` : '';
-    parts.push(`line ${from.x},${from.y}->${to.x},${to.y}${endingText}`);
-  }
-  if (annotation.vertices) {
-    parts.push(`vertices=${annotation.vertices.length}`);
-  }
-  if (annotation.inkPaths) {
-    const pointCount = annotation.inkPaths.reduce((total, path) => total + path.length, 0);
-    parts.push(`inkPaths=${annotation.inkPaths.length}/${pointCount}pts`);
-  }
-  return parts.join('; ');
-}
-
-function visualRegionSources(region: NonNullable<PageResult['visualRegions']>[number]): string {
-  const refs = region.sources.map((source) => `${source.type}[${source.index}]`);
-  const hiddenCount = region.sourceCount - region.sources.length;
-  if (hiddenCount > 0) refs.push(`+${hiddenCount} more`);
-  return refs.join(', ');
-}
-
-function visualRegionAssociatedText(region: NonNullable<PageResult['visualRegions']>[number]): string {
-  return (region.associatedText ?? []).map((item) => `${item.relation}: ${item.text}`).join('; ');
-}
-
-function linkTarget(value: NonNullable<PageResult['links']>[number]['target']): string {
-  return typeof value === 'string' ? value : JSON.stringify(value);
-}
-
-function formatBox(box: { x: number; y: number; width: number; height: number }): string {
-  return `${box.x},${box.y},${box.width},${box.height}`;
-}
-
-function formatBbox(box: number[]): string {
-  return box.join(',');
 }
 
 function layoutBody(page: PageResult, filterRepeated: boolean): string {
@@ -186,29 +66,6 @@ function appendOutline(lines: string[], items: NonNullable<DocumentResult['outli
     lines.push(`${indent}- ${outlineLabel(item)}`);
     if (item.items) appendOutline(lines, item.items, depth + 1);
   }
-}
-
-function formatViewerValue(value: unknown): string {
-  if (typeof value === 'string') return value;
-  if (typeof value === 'number' || typeof value === 'boolean' || value === null) return String(value);
-  return JSON.stringify(value);
-}
-
-function jsActionCount(actions: Record<string, string[]> | undefined): number {
-  return Object.values(actions ?? {}).reduce((sum, scripts) => sum + scripts.length, 0);
-}
-
-function formatJavaScriptActions(actions: Record<string, string[]>): string {
-  const text = Object.entries(actions)
-    .map(([name, scripts]) => `${name}=${scripts.join(' || ')}`)
-    .join(' | ');
-  return truncateForMarkdown(text, MARKDOWN_JS_ACTIONS_MAX_CHARS);
-}
-
-function truncateForMarkdown(text: string, maxChars: number): string {
-  const chars = Array.from(text);
-  if (chars.length <= maxChars) return text;
-  return `${chars.slice(0, maxChars - 3).join('')}...`;
 }
 
 function appendViewer(lines: string[], viewer: NonNullable<DocumentResult['viewer']>): void {
